// src/app/models/Categories.ts
export enum Categories {
    Classici = "Classici",
    NarrativaContemporanea = "Narrativa Contemporanea",
    Fantascienza = "Fantascienza",
    AvventureFantasy = "Avventure/Fantasy",
    GialloThriller = "Giallo/Thriller",
    Biografia = "Biografia",
    Saggistica = "Saggistica",
    RomanzoStorico = "Romanzo Storico",
    Poesia = "Poesia",
    Teatro = "Teatro"
}